Cleanup generator_mateltwise_aarch64
