+++ ide/libxsmm.vcxproj
@@ -67,7 +67,6 @@
     <ClInclude Include="..\src\generator_gemm_sse_avx_avx2_avx512.h" />
     <ClInclude Include="..\src\generator_gemm_sse_microkernel.h" />
     <ClInclude Include="..\src\generator_mateltwise_aarch64.h" />
-    <ClInclude Include="..\src\generator_mateltwise_aarch64_sve.h" />
     <ClInclude Include="..\src\generator_mateltwise_common.h" />
     <ClInclude Include="..\src\generator_mateltwise_gather_scatter_avx_avx512.h" />
     <ClInclude Include="..\src\generator_mateltwise_misc_aarch64.h" />
@@ -147,7 +146,6 @@
     <ClCompile Include="..\src\generator_gemm_sse_microkernel.c" />
     <ClCompile Include="..\src\generator_mateltwise.c" />
     <ClCompile Include="..\src\generator_mateltwise_aarch64.c" />
-    <ClCompile Include="..\src\generator_mateltwise_aarch64_sve.c" />
     <ClCompile Include="..\src\generator_mateltwise_common.c" />
     <ClCompile Include="..\src\generator_mateltwise_gather_scatter_avx_avx512.c" />
     <ClCompile Include="..\src\generator_mateltwise_misc_aarch64.c" />
