+++ ide/libxsmm.vcxproj.filters
@@ -276,9 +276,6 @@
     <ClInclude Include="..\src\generator_mateltwise_gather_scatter_avx_avx512.h">
       <Filter>src\generator</Filter>
     </ClInclude>
-    <ClInclude Include="..\src\generator_mateltwise_aarch64_sve.h">
-      <Filter>src\generator</Filter>
-    </ClInclude>
     <ClInclude Include="..\include\libxsmm_lpflt_quant.h">
       <Filter>include</Filter>
     </ClInclude>
@@ -546,9 +543,6 @@
       <Filter>src\generator</Filter>
     </ClCompile>
     <ClCompile Include="..\src\generator_mateltwise_gather_scatter_avx_avx512.c">
-      <Filter>src\generator</Filter>
-    </ClCompile>
-    <ClCompile Include="..\src\generator_mateltwise_aarch64_sve.c">
       <Filter>src\generator</Filter>
     </ClCompile>
     <ClCompile Include="..\src\libxsmm_lpflt_quant.c">
