+++ src/generator_mateltwise_gather_scatter_aarch64.c
@@ -9,7 +9,6 @@
 /* Evangelos Georganas (Intel Corp.)
 ******************************************************************************/
 #include "generator_mateltwise_aarch64.h"
-#include "generator_mateltwise_aarch64_sve.h"
 #include "generator_aarch64_instructions.h"
 #include "generator_common_aarch64.h"
 #include "generator_common.h"
