+++ src/generator_mateltwise_unary_binary_aarch64.c
@@ -12,7 +12,6 @@
 
 #include "generator_aarch64_instructions.h"
 #include "generator_mateltwise_aarch64.h"
-#include "generator_mateltwise_aarch64_sve.h"
 #include "generator_common_aarch64.h"
 #include "generator_common.h"
 #include "generator_mateltwise_unary_binary_aarch64.h"
